Move out normalizeIDs function to its own lib file
