+++ index.js
@@ -1,6 +1,5 @@
 'use strict'
 
-let Queue = require('promise-queue')
 let RSVP = require('rsvp')
 let _ = require('lodash')
 var fetch = require('./lib/fetch')
@@ -10,6 +9,7 @@ var addInheritedItems = require('./lib/add-inherited-items')
 var putClassesInCouch = require('./lib/classes-in-couch')
 var createVersionIndex = require('./lib/create-version-index')
 let normalizeEmberDependencies = require('./lib/normalize-ember-dependencies')
+let normalizeIDs = require('./lib/normalize-ids')
 let rm = require('rimraf')
 let PouchDB = require('pouchdb')
 require('marked')
@@ -24,11 +24,6 @@ let fs = require('fs')
 
 if (fs.existsSync('tmp/docs')) {
   rm.sync('tmp/docs')
-}
-
-function removeLongDocsBecauseEmber1HasWeirdDocs (document) {
-  let str = 'A Suite can'
-  return document.id.indexOf(str) === -1
 }
 
 function fetchProject (projectName) {
@@ -82,132 +77,3 @@ RSVP.map(projects, fetchProject).then(docs => {
   console.warn('err!', err, err.stack)
   process.exit(1)
 })
-
-function normalizeIDs (versions, projectName) {
-  let tojsonapi = require('yuidoc-to-jsonapi/lib/converter')
-  let updateIDs = require('./lib/update-with-versions-and-project')
-  let findType = require('./lib/filter-jsonapi-doc').byType
-
-  let jsonapidocs = versions.map(version => {
-    let jsonapidoc = tojsonapi(version.data)
-    return updateIDs(jsonapidoc, projectName, version.version)
-  })
-
-  let jsonapidoc = {
-    data: _.flatten(jsonapidocs.map(d => d.data))
-  }
-
-  jsonapidoc.data = jsonapidoc.data.filter(removeLongDocsBecauseEmber1HasWeirdDocs)
-
-  function extractRelationship (doc) {
-    return {
-      id: doc.id,
-      type: doc.type
-    }
-  }
-
-  function filterForVersion (version) {
-    return function (doc) {
-      var projectVersion = doc.relationships['project-version'].data.id.split('-').pop()
-      return version.version === projectVersion
-    }
-  }
-
-  function isPrivate (doc) {
-    return doc.attributes.access === 'private' || doc.attributes.deprecated === true
-  }
-
-  function isPublic (doc) {
-    return doc.attributes.access !== 'private' && doc.attributes.deprecated !== true
-  }
-
-  let projectVersions = versions.map(version => {
-    let classes = findType(jsonapidoc, 'class')
-      .filter(filterForVersion(version))
-      .filter(doc => {
-        return removeLongDocsBecauseEmber1HasWeirdDocs(doc)
-      })
-
-    let namespaces = classes.filter(doc => doc.attributes.static === 1);
-    classes = classes.filter(doc => doc.attributes.static !== 1);
-
-    namespaces.forEach(ns => {
-      ns.type = 'namespace'
-    })
-
-    let modules = findType(jsonapidoc, 'module')
-      .filter(filterForVersion(version))
-
-    return {
-      id: `${projectName}-${version.version}`,
-      type: 'project-version',
-      attributes: {
-        version: version.version
-      },
-      relationships: {
-        classes: {
-          data: classes.map(extractRelationship)
-        },
-        namespaces: {
-          data: namespaces.map(extractRelationship)
-        },
-        modules: {
-          data: modules.map(extractRelationship)
-        },
-        project: {
-          data: {
-            id: projectName,
-            type: 'project'
-          }
-        },
-        'private-classes': {
-          data: classes
-            .filter(isPrivate)
-            .map(extractRelationship)
-        },
-        'public-classes': {
-          data: classes
-            .filter(isPublic)
-            .map(extractRelationship)
-        },
-        'private-namespaces': {
-          data: namespaces
-            .filter(isPrivate)
-            .map(extractRelationship)
-        },
-        'public-namespaces': {
-          data: namespaces
-            .filter(isPublic)
-            .map(extractRelationship)
-        },
-        'private-modules': {
-          data: modules
-            .filter(isPrivate)
-            .map(extractRelationship)
-        },
-        'public-modules': {
-          data: modules
-            .filter(isPublic)
-            .map(extractRelationship)
-        }
-      }
-    }
-  })
-
-  let doc = {
-    data: jsonapidoc.data.concat(projectVersions)
-  }
-
-  let saveDoc = require('./lib/save-document')
-
-  let queue = new Queue(10)
-  let versionDocs = RSVP.map(projectVersions, (projectVersion) => {
-    let doc = {
-      data: projectVersion
-    }
-
-    return queue.add(() => saveDoc(doc))
-  })
-
-  return versionDocs.then(() => doc)
-}
