+++ lib/add-inherited-items.js
@@ -12,13 +12,13 @@ module.exports = function addInheritedItems (docSets) {
 
     classes.forEach(function (klass) {
       let parents = getParents(klass, classes)
-      let thisClassItems = classItems.filter(item => item.class === klass.name)
+      // let thisClassItems = classItems.filter(item => item.class === klass.name)
 
       for (let i = 0; i < parents.length; ++i) {
         parents = parents.concat(getParents(parents[i], classes))
       }
       parents.forEach(function (parent) {
-        if (!parent) return;
+        if (!parent) return
         let parentItems = classItems.filter(item => item.class === parent.name)
         parentItems = parentItems.map(item => {
           item = _.clone(item)
