+++ lib/markup.js
@@ -27,13 +27,12 @@ module.exports = (doc) => {
     replaceDescriptionFor(document.attributes.methods)
     replaceDescriptionFor(document.attributes.properties)
     replaceDescriptionFor(document.attributes.events)
-
   })
 
   return doc
 }
 
-function replaceDescriptionFor(items) {
+function replaceDescriptionFor (items) {
   if (items) {
     items.forEach(item => {
       let itemDescription = item.description
