+++ lib/read-docs.js
@@ -18,7 +18,7 @@ module.exports = function readDocs (docName) {
     } catch (e) {
       console.error(docs)
       console.error(e.stack)
-      process.exit(1);
+      process.exit(1)
     }
 
     return {
